refactor to remove pytest dependecy
