+++ src/reporter/tests/test_1T1E1A.py
@@ -2,7 +2,7 @@ from conftest import QL_URL
 from datetime import datetime
 from exceptions.exceptions import AmbiguousNGSIIdError
 from reporter.tests.utils import insert_test_data, delete_test_data
-from utils.common import assert_equal_time_index_arrays
+from utils.tests.common import assert_equal_time_index_arrays
 import dateutil.parser
 import pytest
 import requests
