+++ src/reporter/tests/test_1T1ENA.py
@@ -2,7 +2,7 @@ from conftest import QL_URL
 from reporter.tests.utils import insert_test_data, delete_test_data
 import pytest
 import requests
-from utils.common import assert_equal_time_index_arrays
+from utils.tests.common import assert_equal_time_index_arrays
 import dateutil.parser
 
 entity_type = 'Room'
