+++ src/reporter/tests/test_1TNENA.py
@@ -1,7 +1,7 @@
 from conftest import QL_URL
 from datetime import datetime
 from reporter.tests.utils import insert_test_data, delete_test_data
-from utils.common import assert_equal_time_index_arrays
+from utils.tests.common import assert_equal_time_index_arrays
 import pytest
 import requests
 import dateutil.parser
